+++ tests/resources/test_iam_role.py
@@ -16,6 +16,13 @@ def iam_role():
                         "Effect": "Allow",
                         "Principal": {"Service": ["ec2.amazonaws.com"], "AWS": "arn:aws:iam::111111111111:root"},
                         "Action": ["sts:AssumeRole"],
+                        "Condition": {
+                            "StringLike": {
+                                "iam:AssociatedResourceARN": [
+                                    "arn:aws:ec2:us-east-1:999999999999:instance/*",
+                                ]
+                            },
+                        },
                     },
                 },
                 "Path": "/",
@@ -24,7 +31,20 @@ def iam_role():
                         "PolicyName": "root",
                         "PolicyDocument": {
                             "Version": "2012-10-17",
-                            "Statement": {"Effect": "Allow", "Action": "*", "Resource": "*"},
+                            "Statement": {
+                                "Effect": "Allow",
+                                "Action": "*",
+                                "Resource": "*",
+                                "Condition": {
+                                    "StringEquals": {"iam:PassedToService": "ec2.amazonaws.com"},
+                                    "StringLike": {
+                                        "iam:AssociatedResourceARN": [
+                                            "arn:aws:ec2:us-east-1:111122223333:instance/*",
+                                            "arn:aws:ec2:us-west-1:111122223333:instance/*",
+                                        ]
+                                    },
+                                },
+                            },
                         },
                     }
                 ],
@@ -37,6 +57,22 @@ def test_policies(iam_role):
     policies = iam_role.Properties.Policies
     assert len(policies) == 1
     assert policies[0].PolicyName == "root"
+
+
+def test_all_conditions(iam_role):
+    assert iam_role.all_statement_conditions[0].StringEquals == {"iam:PassedToService": "ec2.amazonaws.com"}
+    assert iam_role.all_statement_conditions[0].StringLike == {
+        "iam:AssociatedResourceARN": [
+            "arn:aws:ec2:us-east-1:111122223333:instance/*",
+            "arn:aws:ec2:us-west-1:111122223333:instance/*",
+        ]
+    }
+
+    assert iam_role.assume_role_statement_conditions[0].StringLike == {
+        "iam:AssociatedResourceARN": [
+            "arn:aws:ec2:us-east-1:999999999999:instance/*",
+        ]
+    }
 
 
 def test_iamrole_policy_documents(iam_role):